Implemented subsidy halving block size limiting.

+++ include/cli.h
@@ -14,7 +14,7 @@ class CLI {
         void listAddresses();
         void printChain();
         void reindexUTXO();
-        void send(const std::string& from, const std::string& to, int amount);
+        void send(const std::string& from, const std::string& to, int64_t amount);
         void startNode(uint16_t port, const std::string& seedAddr, uint16_t rpcPort,
                        const std::string& minerAddress);
